refactor: split task scheduler into independent schedulers
